requirements.txt
faiss-cpu>=1.8.0
modelscope>=1.14.0
numpy>=1.24.0
requests>=2.31.0
torch>=2.1.0
transformers>=4.39.0
tqdm>=4.66.0
rank-bm25>=0.2.2

